Fixed sodium shader recompile spam

+++ mods/Veil/neoforge/src/main/java/foundry/veil/forge/mixin/compat/sodium/ShaderChunkRendererMixin.java
@@ -2,6 +2,7 @@ package foundry.veil.forge.mixin.compat.sodium;
 
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.mojang.blaze3d.systems.RenderSystem;
 import foundry.veil.Veil;
 import foundry.veil.api.client.render.VeilRenderSystem;
 import foundry.veil.forge.ext.ShaderChunkRendererExtension;
@@ -13,7 +14,6 @@ import net.caffeinemc.mods.sodium.client.gl.shader.*;
 import net.caffeinemc.mods.sodium.client.render.chunk.ShaderChunkRenderer;
 import net.caffeinemc.mods.sodium.client.render.chunk.shader.ChunkShaderInterface;
 import net.caffeinemc.mods.sodium.client.render.chunk.shader.ChunkShaderOptions;
-import net.minecraft.Util;
 import net.minecraft.client.Minecraft;
 import net.minecraft.resources.ResourceLocation;
 import org.spongepowered.asm.mixin.Final;
@@ -24,9 +24,10 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-import java.util.*;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.Collections;
+import java.util.Map;
+import java.util.Queue;
+import java.util.concurrent.ConcurrentLinkedDeque;
 
 import static org.lwjgl.opengl.GL20C.GL_FRAGMENT_SHADER;
 import static org.lwjgl.opengl.GL20C.GL_VERTEX_SHADER;
@@ -42,16 +43,14 @@ public abstract class ShaderChunkRendererMixin implements ShaderChunkRendererExt
     protected abstract GlProgram<ChunkShaderInterface> createShader(String path, ChunkShaderOptions options);
 
     @Unique
-    private CompletableFuture<?> veil$sources;
-    @Unique
     private ThreadTaskScheduler scheduler;
     @Unique
     private Map<ShaderType, String> veil$shaderSource;
 
     @Inject(method = "delete", at = @At("HEAD"), remap = false)
     public void delete(CallbackInfo ci) {
-        if (this.veil$sources != null) {
-            this.veil$sources.cancel(false);
+        if (this.scheduler != null) {
+            this.scheduler.cancel();
         }
     }
 
@@ -68,52 +67,56 @@ public abstract class ShaderChunkRendererMixin implements ShaderChunkRendererExt
 
     @Override
     public void veil$recompile(int activeBuffers) {
-        if (this.veil$sources != null) {
-            this.veil$sources.cancel(false);
+        if (this.scheduler != null) {
+            this.scheduler.cancel();
         }
 
-        List<ChunkShaderOptions> keys = new ArrayList<>(this.programs.keySet());
-        this.veil$sources = CompletableFuture.supplyAsync(() -> {
-            Map<ChunkShaderOptions, Map<ShaderType, String>> sources = new HashMap<>(2);
-            ResourceLocation vsh = ResourceLocation.fromNamespaceAndPath("sodium", "blocks/block_layer_opaque.vsh");
-            ResourceLocation fsh = ResourceLocation.fromNamespaceAndPath("sodium", "blocks/block_layer_opaque.fsh");
-            Map<ResourceLocation, String> shaders = Map.of(
-                    vsh,
-                    ShaderLoader.getShaderSource(vsh),
-                    fsh,
-                    ShaderLoader.getShaderSource(fsh)
-            );
+        ResourceLocation vsh = ResourceLocation.fromNamespaceAndPath("sodium", "blocks/block_layer_opaque.vsh");
+        ResourceLocation fsh = ResourceLocation.fromNamespaceAndPath("sodium", "blocks/block_layer_opaque.fsh");
+        Map<ResourceLocation, String> shaders = Map.of(
+                vsh,
+                ShaderLoader.getShaderSource(vsh),
+                fsh,
+                ShaderLoader.getShaderSource(fsh)
+        );
 
-            SimpleShaderProcessor.setup(Minecraft.getInstance().getResourceManager(), activeBuffers, Collections.singleton(new SodiumShaderProcessor()));
-            for (ChunkShaderOptions option : keys) {
+        Queue<ChunkShaderOptions> keys = new ConcurrentLinkedDeque<>(this.programs.keySet());
+        int shaderCount = keys.size();
+        this.scheduler = new ThreadTaskScheduler("VeilSodiumShaderCompile", 1, () -> {
+            ChunkShaderOptions option = keys.poll();
+            if (option == null) {
+                return null;
+            }
+            return () -> {
                 Map<ShaderType, String> map = new Object2ObjectArrayMap<>();
                 for (Map.Entry<ResourceLocation, String> entry : shaders.entrySet()) {
                     ResourceLocation shader = entry.getKey();
                     String src = ShaderParser.parseShader(entry.getValue(), option.constants());
                     boolean vertex = shader.getPath().endsWith(".vsh");
                     try {
+                        SimpleShaderProcessor.setup(Minecraft.getInstance().getResourceManager(), activeBuffers, Collections.singleton(new SodiumShaderProcessor()));
                         src = SimpleShaderProcessor.modify(null, ResourceLocation.fromNamespaceAndPath(shader.getNamespace(), "shaders/" + shader.getPath()), null, vertex ? GL_VERTEX_SHADER : GL_FRAGMENT_SHADER, src);
+                        SimpleShaderProcessor.free();
                     } catch (Exception e) {
                         Veil.LOGGER.error("Failed to apply Veil shader modifiers to shader: {}", shader, e);
                     }
                     map.put(vertex ? ShaderType.VERTEX : ShaderType.FRAGMENT, src);
                 }
-                sources.put(option, map);
-            }
-            SimpleShaderProcessor.free();
-
-            return sources;
-        }, Util.backgroundExecutor()).thenAcceptAsync(sources -> {
-            for (Map.Entry<ChunkShaderOptions, Map<ShaderType, String>> entry : sources.entrySet()) {
-                ChunkShaderOptions options = entry.getKey();
-                this.veil$shaderSource = entry.getValue();
-                GlProgram<ChunkShaderInterface> old = this.programs.put(options, this.createShader("blocks/block_layer_opaque", options));
-                if (old != null) {
-                    old.delete();
-                }
-            }
+                RenderSystem.recordRenderCall(() -> {
+                    this.veil$shaderSource = map;
+                    GlProgram<ChunkShaderInterface> old = this.programs.put(option, this.createShader("blocks/block_layer_opaque", option));
+                    if (old != null) {
+                        old.delete();
+                    }
+                    this.veil$shaderSource = null;
+                });
+            };
+        });
+        this.scheduler.getCompletedFuture().thenRunAsync(() -> {
             this.veil$shaderSource = null;
-            Veil.LOGGER.info("Recompiled {} Sodium Shaders", sources.values().stream().mapToInt(Map::size).sum());
+            if (!this.scheduler.isCancelled()) {
+                Veil.LOGGER.info("Recompiled {} Sodium Shaders", shaderCount);
+            }
         }, VeilRenderSystem.renderThreadExecutor());
     }
